nucypher.config.node.CharacterConfiguration --> nucypher.config.characters.CharacterConfiguration
